+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 This project changes will be shown here.
 
+## [unreleased]
+### Added
+- Issue 967 - Record half day holidays to assign logged hours
+
 ## [v4.5.4] - 2020-09-02
 ### Added
 - Add the ability to add items to Credit Notes
-
 
 ## [v4.5.3] - 2020-09-01
 ### Fixed